organise files abit
