+++ application/imports.py
@@ -9,7 +9,25 @@ import uuid
 
 
 # Imports from other files
-from login_service import Login_Service
+
+# controllers
+from classes.controllers.actionscontroller import ActionsController
+
+# infosec services
+from classes.infosec.login_service import Login_Service
+from classes.infosec.encryption_service import Encryption_Service
+from classes.infosec.authorisation_service import Authorisation_Service
+from classes.infosec.input_sanitisation import Input_Sanitisation_Service
+
+# logging services
+from classes.logger.logger import Logger
+from classes.logger.auditor import Auditor
+
+# sensor components
+from classes.sensors.geigercounter import GeigerCounter
+from classes.sensors.thermometer import Thermometer
+
+# misc
+from classes.authseeder import AuthSeeder
 from classes.dbmanager import DBManager
-from input_sanitisation import Input_Sanitisation_Service
-from user import User
+from classes.commandline_interface import CommandLineInterface
