+++ application/tests/unit/context.py
@@ -14,7 +14,7 @@ import sys
 path = Path(__file__).resolve().parent.parent.parent
 sys.path.insert(0, str(path))
 
-from commandline_interface import CommandLineInterface
+from classes.commandline_interface import CommandLineInterface
 from classes.sensors.geigercounter import GeigerCounter
 from classes.sensors.thermometer import Thermometer
-from classes.authorisation_service import AuthorisationService
+from classes.infosec.authorisation_service import AuthorisationService
